Fix import path and rename file
