+++ web/packages/teleport/src/Discover/Server/CreateEc2Ice/CreateEc2Ice.tsx
@@ -21,7 +21,7 @@ import { Danger } from 'design/Alert';
 import { FetchStatus } from 'design/DataTable/types';
 
 import useAttempt from 'shared/hooks/useAttemptNext';
-import { getErrMessage } from '@gravitational/shared/utils/errorType';
+import { getErrMessage } from 'shared/utils/errorType';
 
 import {
   SecurityGroup,
